+++ include/slepc/private/rgimpl.h
@@ -23,6 +23,7 @@ struct _RGOps {
   PetscErrorCode (*istrivial)(RG,PetscBool*);
   PetscErrorCode (*computecontour)(RG,PetscInt,PetscScalar*,PetscScalar*);
   PetscErrorCode (*computebbox)(RG,PetscReal*,PetscReal*,PetscReal*,PetscReal*);
+  PetscErrorCode (*computequadrature)(RG,RGQuadRule,PetscInt,PetscScalar*,PetscScalar*,PetscScalar*);
   PetscErrorCode (*checkinside)(RG,PetscReal,PetscReal,PetscInt*);
   PetscErrorCode (*setfromoptions)(PetscOptionItems*,RG);
   PetscErrorCode (*view)(RG,PetscViewer);