+++ src/sys/classes/rg/impls/ellipse/rgellipse.c
@@ -200,6 +200,28 @@ PetscErrorCode RGComputeBoundingBox_Ellipse(RG rg,PetscReal *a,PetscReal *b,Pets
   PetscFunctionReturn(0);
 }
 
+PetscErrorCode RGComputeQuadrature_Ellipse(RG rg,RGQuadRule quad,PetscInt n,PetscScalar *z,PetscScalar *zn,PetscScalar *w)
+{
+  RG_ELLIPSE *ctx = (RG_ELLIPSE*)rg->data;
+  PetscReal  theta;
+  PetscInt   i;
+
+  PetscFunctionBegin;
+  for (i=0;i<n;i++) {
+#if defined(PETSC_USE_COMPLEX)
+    theta = 2.0*PETSC_PI*(i+0.5)/n;
+    zn[i] = PetscCMPLX(PetscCosReal(theta),ctx->vscale*PetscSinReal(theta));
+    w[i]  = rg->sfactor*ctx->radius*(PetscCMPLX(ctx->vscale*PetscCosReal(theta),PetscSinReal(theta)))/n;
+#else
+    theta = PETSC_PI*(i+0.5)/n;
+    zn[i] = PetscCosReal(theta);
+    w[i]  = PetscCosReal((n-1)*theta)/n;
+    z[i]  = rg->sfactor*(ctx->center + ctx->radius*zn[i]);
+#endif
+  }
+  PetscFunctionReturn(0);
+}
+
 PetscErrorCode RGCheckInside_Ellipse(RG rg,PetscReal px,PetscReal py,PetscInt *inside)
 {
   RG_ELLIPSE *ctx = (RG_ELLIPSE*)rg->data;
@@ -261,13 +283,14 @@ SLEPC_EXTERN PetscErrorCode RGCreate_Ellipse(RG rg)
   ellipse->vscale = 1.0;
   rg->data = (void*)ellipse;
 
-  rg->ops->istrivial      = RGIsTrivial_Ellipse;
-  rg->ops->computecontour = RGComputeContour_Ellipse;
-  rg->ops->computebbox    = RGComputeBoundingBox_Ellipse;
-  rg->ops->checkinside    = RGCheckInside_Ellipse;
-  rg->ops->setfromoptions = RGSetFromOptions_Ellipse;
-  rg->ops->view           = RGView_Ellipse;
-  rg->ops->destroy        = RGDestroy_Ellipse;
+  rg->ops->istrivial         = RGIsTrivial_Ellipse;
+  rg->ops->computecontour    = RGComputeContour_Ellipse;
+  rg->ops->computebbox       = RGComputeBoundingBox_Ellipse;
+  rg->ops->computequadrature = RGComputeQuadrature_Ellipse;
+  rg->ops->checkinside       = RGCheckInside_Ellipse;
+  rg->ops->setfromoptions    = RGSetFromOptions_Ellipse;
+  rg->ops->view              = RGView_Ellipse;
+  rg->ops->destroy           = RGDestroy_Ellipse;
   ierr = PetscObjectComposeFunction((PetscObject)rg,"RGEllipseSetParameters_C",RGEllipseSetParameters_Ellipse);CHKERRQ(ierr);
   ierr = PetscObjectComposeFunction((PetscObject)rg,"RGEllipseGetParameters_C",RGEllipseGetParameters_Ellipse);CHKERRQ(ierr);
   PetscFunctionReturn(0);