+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - New `DSGSVD` type for the generalized SVD.
 - New contour integral method in `DSNEP`.
 - New function `BVSetRandomSign()` for random values equal to -1 or 1 with equal probability.
+- New function `RGComputeQuadrature()` to compute quadrature values used in contour
+  integral solvers.
 
 ### Changed